custom-elements/concepts/custom-elements-type-test-006.html

<!DOCTYPE html>
<html>
<head>
<title>The custom element type started with combining char</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="The custom element type cannot start with combining char">
<link rel="help" href="https://dvcs.w3.org/hg/webcomponents/raw-file/default/spec/custom/index.html#concepts">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

function doTest(charCode) {
    var doc = newHTMLDocument();
    
    var XFoo = null;
    
    var thrown = false;    
    var name = String.fromCharCode(charCode) + '-foo';
    try {
        XFoo = doc.register(name);
    } catch (e) {
        thrown = true;
    }
    
    assert_true(thrown, 'Exception should be thrown in case of attempt to register '
            + 'a custom element with a name \'' + name + '\'  (char code U+' + charCode.toString(16) + ' )');
}

var parameters = [];

for (var i = 0; i < COMBINING_CHARS.length; i++) {
    parameters[i] = ['Custom element name ' + String.fromCharCode(COMBINING_CHARS[i]) + '-foo', 
                     COMBINING_CHARS[i]];
}

generate_tests(doTest, parameters, 'Invalid custom element types started with combining char');

</script>
</body>
</html>
